Add Mutex in GarbageCollectionProfiler.total_time
When the internal garbage collection profiler is slow to clear the total
GC time, Appsignal::GarbageCollectionProfiler#total_time uses a Mutex to
lock and make sure GC time is not counted twice.

The test for this sets a clear delay, which is used by FakeGCProfiler to
pause between returning the total time and clearing it. The test sets a
total time for the first thread, which gets cleared and should be 0 for
the second one.